contracts/lib/SafeMath: mul64 fixed

+++ contracts/lib/SafeMath.sol
@@ -66,7 +66,7 @@ library SafeMath {
   /**
   * @dev Multiplies two numbers, reverts on overflow.
   */
-  function mul64(uint64 a, uint64 b) internal pure returns (uint256) {
+  function mul64(uint64 a, uint64 b) internal pure returns (uint64) {
     // Gas optimization: this is cheaper than requiring 'a' not being zero, but the
     // benefit is lost if 'b' is also tested.
     // See: https://github.com/OpenZeppelin/openzeppelin-solidity/pull/522
@@ -74,7 +74,7 @@ library SafeMath {
       return 0;
     }
 
-    uint256 c = a * b;
+    uint64 c = a * b;
     require(c / a == b);
 
     return c;
